+++ src/Phan/Config.php
@@ -420,6 +420,9 @@ class Config
         // As a result, enabling this setting with target_php_version 8.0 may result in false positives for `--redundant-condition-detection` when codebases also support php 7.x.
         'assume_real_types_for_internal_functions' => false,
 
+        // If enabled, Phan will use the php 8.1+ tentative return types available for PHP and extensions.
+        'use_tentative_return_type' => true,
+
         // If enabled, scalars (int, float, bool, string, null)
         // are treated as if they can cast to each other.
         // This does not affect checks of array keys. See `scalar_array_key_cast`.