+++ tests/Phan/Internal/ConfigEntry.php
@@ -69,6 +69,7 @@ class ConfigEntry
         'assume_no_external_class_overrides' => self::CATEGORY_ANALYSIS,
         'allow_method_param_type_widening' => self::CATEGORY_ANALYSIS_VERSION,
         'guess_unknown_parameter_type_using_default' => self::CATEGORY_ANALYSIS,
+        'use_tentative_return_type' => self::CATEGORY_ANALYSIS,
         'allow_overriding_vague_return_types' => self::CATEGORY_ANALYSIS,
         'infer_default_properties_in_construct' => self::CATEGORY_ANALYSIS,
         'inherit_phpdoc_types' => self::CATEGORY_ANALYSIS,